SERVER-3719 long long values in v1 indexes were losing precision

+++ db/key.cpp
@@ -305,7 +305,7 @@ namespace mongo {
                 {
                     long long n = e._numberLong();
                     double d = (double) n;
-                    if( d != n ) { 
+                    if( ((long long)d) != n ) { 
                         traditional(obj);
                         return;
                     }
